Add Sample Flask App
[Problem]
It seems like Flask will be a decent framework to
run this application on:
* Can run on EC2 later if I decide lambda isn't up to task
* Pretty well-documented
* Fits desired language of Python

[Solution]
Put together an app using a guide. It seems I can add routes
pretty easily. Downside is that there's not room to deploy
separate lambda functions per API.

Luckily, my goal for this service is to be insanely lightweight,
so it's not a giant concern.

Will likely move to EC2 once I'm out of the early testing phases
(don't want to pay for an instance).